Moved files to more fitting folders
